updated more types
